antispoofing-backend/app/requirements.txt
Django
djangorestframework
markdown
django-filter
psycopg2-binary
gunicorn
django-storages
django-environ
boto3
confluent-kafka
django-cors-headers